refactor(pty): split ProcessDetector into domain modules
Extract free functions, types, registries, and command-parsing into
separate modules under ProcessDetector/ while preserving the original
ProcessDetector.ts as a barrel re-export. Consolidate duplicated
logIdentityDebug into shared pty/identityDebug.ts.

Modules:
- ProcessDetector/types.ts — interfaces, DetectionResult, factory fns
- ProcessDetector/registries.ts — AGENT_CLI_NAMES, PROCESS_ICON_MAP, etc.
- ProcessDetector/commandParser.ts — extractCommandNameCandidates, etc.
- ProcessDetector/candidateHelpers.ts — buildDetectedCandidate, etc.
- ProcessDetector/ProcessDetector.ts — class (hysteresis, lifecycle)
- pty/identityDebug.ts — shared debug logger (was duplicated)

No consumer import paths change. 139 tests pass. #6692
