+++ electron/services/ProcessDetector/types.ts
@@ -1,0 +1,108 @@
+import type { BuiltInAgentId } from "../../../shared/config/agentIds.js";
+
+export interface ChildProcess {
+  pid: number;
+  name: string;
+  command?: string;
+}
+
+export interface DetectedProcessCandidate {
+  agentType?: BuiltInAgentId;
+  processIconId?: string;
+  processName: string;
+  processCommand?: string;
+  priority: number;
+  order: number;
+}
+
+export interface CommandIdentity {
+  agentType?: BuiltInAgentId;
+  processIconId?: string;
+  processName: string;
+}
+
+/**
+ * Explicit detection state. Ambiguity is first-class — `unknown` means we have
+ * no evidence (cache error, blind `ps`, invalid PID) and should not mutate
+ * committed state; `ambiguous` means we have conflicting positive evidence
+ * from two independent sources and are holding until one stabilises. Only
+ * `agent` and `no_agent` drive actual state changes in consumers.
+ */
+export type DetectionState = "unknown" | "no_agent" | "agent" | "ambiguous";
+
+/** Which signal produced the committed agent identity, for diagnostics. */
+export type DetectionEvidenceSource = "process_tree" | "shell_command" | "both";
+
+export interface DetectionResult {
+  detectionState: DetectionState;
+  /** @deprecated Use `detectionState === "agent"`. Retained for legacy consumers. */
+  detected: boolean;
+  agentType?: BuiltInAgentId;
+  processIconId?: string;
+  processName?: string;
+  isBusy?: boolean;
+  currentCommand?: string;
+  evidenceSource?: DetectionEvidenceSource;
+}
+
+export type DetectionCallback = (result: DetectionResult, spawnedAt: number) => void;
+
+export function makeAgentResult(params: {
+  agentType?: BuiltInAgentId;
+  processIconId?: string;
+  processName?: string;
+  isBusy?: boolean;
+  currentCommand?: string;
+  evidenceSource?: DetectionEvidenceSource;
+}): DetectionResult {
+  return {
+    detectionState: "agent",
+    detected: true,
+    agentType: params.agentType,
+    processIconId: params.processIconId,
+    processName: params.processName,
+    isBusy: params.isBusy,
+    currentCommand: params.currentCommand,
+    evidenceSource: params.evidenceSource,
+  };
+}
+
+export function makeNoAgentResult(params: {
+  isBusy?: boolean;
+  currentCommand?: string;
+  evidenceSource?: DetectionEvidenceSource;
+}): DetectionResult {
+  return {
+    detectionState: "no_agent",
+    detected: false,
+    isBusy: params.isBusy,
+    currentCommand: params.currentCommand,
+    evidenceSource: params.evidenceSource,
+  };
+}
+
+export function makeUnknownResult(params?: {
+  isBusy?: boolean;
+  currentCommand?: string;
+}): DetectionResult {
+  return {
+    detectionState: "unknown",
+    detected: false,
+    isBusy: params?.isBusy,
+    currentCommand: params?.currentCommand,
+  };
+}
+
+export function makeAmbiguousResult(params: {
+  isBusy?: boolean;
+  currentCommand?: string;
+  evidenceSource?: DetectionEvidenceSource;
+}): DetectionResult {
+  return {
+    detectionState: "ambiguous",
+    detected: false,
+    isBusy: params.isBusy,
+    currentCommand: params.currentCommand,
+    evidenceSource: params.evidenceSource,
+  };
+}
